All changes made since April 12

+++ packages/buttons-frontend/src/App.js
@@ -1,16 +1,21 @@
 import React, { Component } from 'react';
 import './App.css';
-import Members from './components/memberList/Members';
-import Locations from './components/locationList/Locations';
+import Users from './components/homePage/userList/Users';
+import Locations from './components/homePage/locationList/Locations';
 import AllLocations from './components/adminEdit/editLocations/AllLocations';
-import AllMembers from './components/adminEdit/editUsers/AllMembers';
+import AllUsers from './components/adminEdit/editUsers/AllUsers';
 import Navigation from './components/navigation/Navigation';
 import AddUser from './components/adminEdit/addUser/AddUser';
 import AddLocation from './components/adminEdit/addLocation/AddLocation';
 import Error from './components/error/Error';
+import UsersAlt from './components/homePage/usersNotAdmin/UsersAlt';
+import NoAccess from './components/adminEdit/noAccess/NoAccess';
+import Logout from './components/logout/Logout';
 import Cookies from 'js-cookie';
 import styles from './App.module.css';
 import { BrowserRouter, Route, Switch} from 'react-router-dom';
+
+// Various functions that change user/location database info
 
 function getSession() {
   return JSON.parse(atob(Cookies.get('koa:sess')))
@@ -24,7 +29,12 @@ async function getUsers () {
         'X-CSRF-Token': getSession().csrfToken
     }
   })
+  try {
   return (await res.json())
+}
+  catch(err) {
+    console.log(err + 'Error with get Users')
+  }
 }
 
 async function getLocations () {
@@ -51,23 +61,18 @@ async function changeLoc(user) {
     body: JSON.stringify({
       LocationId: user.locId
     })
-    //how should I also make this update the state as well?
 })
 }
 
 
 
 async function leaveLoc(user) {
-  const url = '/api/users/' + user.userId
+  const url = '/api/users/' + user.userId + '/LocationId'
   const res = await fetch(url, {
-    method: 'PATCH',
+    method: 'DELETE',
     headers: {
-      'Content-Type': 'application/json',
       'X-CSRF-Token': getSession().csrfToken
     },
-    body: JSON.stringify({
-      LocationId: null
-      })
     })
 }
 
@@ -86,7 +91,7 @@ async function changeLocData(location) {
     })
 }
 
-async function changeMemberData(user) {
+async function changeUserData(user) {
   console.log(user.name)
   const url = '/api/users/' + user.id
   const res = await fetch(url, {
@@ -139,14 +144,17 @@ async function addLocation(location) {
 }
 
 async function deleteUser(user) {
-  const url = '/api/users/' + user.id
-  const res = await fetch(url, {
-    method: 'DELETE',
-    headers: {
-      'X-CSRF-Token': getSession().csrfToken
-    }
-    })
-  return (await res.json())
+  const result = window.confirm('Are you sure you want to delete this user?')
+  if (result) {
+    const url = '/api/users/' + user.id
+    const res = await fetch(url, {
+      method: 'DELETE',
+      headers: {
+        'X-CSRF-Token': getSession().csrfToken
+      }
+      })
+    return (await res.json())
+  }
 }
 
 async function deleteLoc(loc) {
@@ -160,20 +168,29 @@ async function deleteLoc(loc) {
   return (await res.json())
 }
 
+function logoutFunc() {
+  Cookies.remove('koa:sess', {path: '/'});
+}
+window.Cookies = Cookies
+
+
+//Class App begins
 
 class App extends Component {
   state = {
-    members: [],
-    locations: []
+    users: [],
+    locations: [],
+    loaded: false
   }
 
   async componentDidMount () {
-    this.setState({locations: await getLocations(), members: await getUsers()})
+    this.setState({locations: await getLocations(), users: await getUsers(), loaded: true})
   }
 
+//Changes state directly and calls db function
   changeLocHelp = (user) => {
     {changeLoc(user)}
-    this.setState({members: this.state.members.map(member => {
+    this.setState({users: this.state.users.map(member => {
       if (member.id === user.userId) {
         member.LocationId = user.locId
       }
@@ -182,49 +199,87 @@ class App extends Component {
   })
   }
 
+//Changes state directly and calls db function
 leaveLocHelp = (user) => {
   {leaveLoc(user)}
-  this.setState({members: this.state.members.map(member => {
+  this.setState({users: this.state.users.map(member => {
     if (member.id === user.userId) {
-      member.LocationId = 0
+      member.LocationId = null
     }
     return member
   })
 })
 }
 
+
+//Page specifications
+//if the user is an admin, the home page will look different than if they are not
   HomePage = () => {
-    return (
-      <div className = {styles['everything']}>
-        <h1>Where To Work</h1>
-        <div className = {styles['membersLocationsContainer']}>
-          <Members
-            locations = {this.state.locations}
-            members = {this.state.members}
-            changeLocHelp = {this.changeLocHelp}
-          />
-          <Locations
-            members = {this.state.members}
-            locations = {this.state.locations}
-            leaveLoc = {leaveLoc}
-            leaveLocHelp = {this.leaveLocHelp}
-          />
+    if (((JSON.parse(atob(Cookies.get('koa:sess')))).user.administrator) === true) {
+      return (
+        <div>
+          <h1>Where To Work</h1>
+          <div className = {styles['usersLocationsContainer']}>
+            <Users
+              locations = {this.state.locations}
+              users = {this.state.users}
+              changeLocHelp = {this.changeLocHelp}
+            />
+            <Locations
+              users = {this.state.users}
+              locations = {this.state.locations}
+              leaveLoc = {leaveLoc}
+              leaveLocHelp = {this.leaveLocHelp}
+            />
+          </div>
         </div>
-      </div>
-      )
+        )
+      } else {
+        return (
+          <div>
+            <h1>Where to Work</h1>
+            <div className= {styles['usersLocationsContainer']}>
+              <UsersAlt
+                locations = {this.state.locations}
+                users = {this.state.users}
+                changeLocHelp = {this.changeLocHelp}
+              />
+              <Locations
+                users = {this.state.users}
+                locations = {this.state.locations}
+                leaveLoc = {leaveLoc}
+                leaveLocHelp = {this.leaveLocHelp}
+              />
+            </div>
+          </div>
+          )
+      }
   }
 
   Error = () => {
     return (
-      <div className = {styles['everything']}>
+      <div>
         <Error/>
       </div>
       )
   }
 
-  AdminPage = () => {
+  Logout = () => {
     return (
-      <div className = {styles['everything']}>
+      <Logout
+        logoutFunc = {logoutFunc}
+      />
+      )
+  }
+
+  AdminPage = () => {
+    if (((JSON.parse(atob(Cookies.get('koa:sess')))).user.administrator) !== true) {
+      return (
+        <NoAccess/>
+        )
+    }
+    return (
+      <div className>
         <AddUser
           addUser = {addUser}
         />
@@ -236,9 +291,9 @@ leaveLocHelp = (user) => {
           changeLocData = {changeLocData}
           deleteLoc = {deleteLoc}
         />
-        <AllMembers
-          members = {this.state.members}
-          changeMemberData = {changeMemberData}
+        <AllUsers
+          users = {this.state.users}
+          changeUserData = {changeUserData}
           deleteUser = {deleteUser}
         />
       </div>
@@ -246,18 +301,28 @@ leaveLocHelp = (user) => {
   }
 
   render () {
-    return (
-      <BrowserRouter>
-        <div className = {styles['background']}>
-          <Navigation />
-          <Switch>
-            <Route path='/' component = {this.HomePage} exact />
-            <Route path='/admin' component = {this.AdminPage} />
-            <Route component={this.Error} />
-          </Switch>
+    if (this.state.loaded === false) {
+      return (
+        <div>
+          <div className = {styles['loadingItem']}></div>
+          <div className = {styles['loadingItem2']}></div>
         </div>
-      </BrowserRouter>
-    )
+        )
+    } else {
+      return (
+        <BrowserRouter>
+          <div className = {styles['background']}>
+            <Navigation />
+            <Switch>
+              <Route path='/' component = {this.HomePage} exact />
+              <Route path='/admin' component = {this.AdminPage} />
+              <Route path ='/logout' component={this.Logout}/>
+              <Route component={this.Error} />
+            </Switch>
+          </div>
+        </BrowserRouter>
+      )
+    }
   }
 }
